ccs/pyproject.toml
[project]
name = "ccs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
